Finish refactor and changes for experiment 1.
This commit addresses the concerns around how participants are
tied to one another. Before, ties were getting double-counted
in many cases, due to how data was represented in Breadboard. The
commit deals with this and also addresses destroying ties when
someone leaves an experiment. Additionally, it drops experiments
that don't meet certain criteria in terms of number of players or
rounds.