Add a whole bunch of useful files for processing flash mobs.
This commit is the culmination of many scripts written in a one-off
manner being put to better use to generate the needed files to
carry out a flash mob. The file `setup_flash_mob_files.sh` in
particular is important as it takes relevant python scripts and
strings them together to create predictable names and outputs for
further use.